MIDI in: Output frequency control voltage, active note stack
Pre-calculates all the values needed to map available notes to output
values that will map to 1V/Oct voltages.

The note stack allows up to 24 (well, actually 25) notes to be held at a
time, letting you release the latest note to return to the previously
pressed note, as long as it's still held down.

Further, this documents the code a bit more.