Add indication LEDs for what displayed note means
There are 3 cases a note is displayed on the screen: when pressing a
step when not running, when changing the max note, when changing the min
note.

These LEDs light up next to a label, indicating what the displayed note
means. They use the 7 segment driver.

Further: cleaned up all int types in the sequencer codebase to use
stdint ones, specifying exact size.